+++ connectors/kafka/kafka-connector/src/main/java/org/platformlambda/kafka/ServiceRegistry.java
@@ -50,6 +50,7 @@ public class ServiceRegistry implements LambdaFunction {
     private static final String LEAVE = "leave";
     private static final String RESTART = "restart";
     private static final String CHECKSUM = "checksum";
+    private static final String TARGET = "target";
     private static final String PING = "ping";
     private static final String STOP = "stop";
     // static because this is a shared lambda function
@@ -144,7 +145,7 @@ public class ServiceRegistry implements LambdaFunction {
     @SuppressWarnings("unchecked")
     public Object handleEvent(Map<String, String> headers, Object body, int instance) throws IOException {
         if (isServiceMonitor) {
-            // service monitor does not use global routing table
+            // ignore requests since service monitor does not maintain routing table
             return false;
         }
         String type = headers.get(TYPE);
@@ -186,7 +187,6 @@ public class ServiceRegistry implements LambdaFunction {
         String type = headers.get(TYPE);
         if (PING.equals(type)) {
             broadcastChecksum(getChecksum());
-            log.info("Routing table integrity check");
             return true;
         }
         if (CHECKSUM.equals(type) && headers.containsKey(CHECKSUM) && headers.containsKey(ORIGIN)) {
@@ -196,10 +196,11 @@ public class ServiceRegistry implements LambdaFunction {
             String myChecksum = getChecksum();
             if (!origin.equals(myOrigin)) {
                 if (extChecksum.equals(myChecksum)) {
-                    log.debug("Routing table matches with {}, checksum {}", origin, extChecksum);
+                    log.debug("Routing table matches with {}, checksum={}", origin, extChecksum);
                 } else {
                     log.warn("Routing table checksum not matched. Syncing with {}", origin);
                     sendMyRoutes(origin);
+                    downloadFromPeer(origin);
                 }
             }
         }
@@ -316,16 +317,32 @@ public class ServiceRegistry implements LambdaFunction {
         }
     }
 
+    private void downloadFromPeer(String peer) throws IOException {
+        String origin = Platform.getInstance().getOrigin();
+        EventEnvelope request = new EventEnvelope();
+        request.setTo(ServiceDiscovery.SERVICE_REGISTRY + "@" + peer)
+                .setHeader(TYPE, JOIN).setHeader(ORIGIN, origin);
+        origins.put(peer, Utility.getInstance().date2str(new Date(), true));
+        PostOffice.getInstance().send(request);
+    }
+
     private void broadcastChecksum(String checkSum) throws IOException {
         PostOffice po = PostOffice.getInstance();
-        String myOrigin = Platform.getInstance().getOrigin();
+        String origin = Platform.getInstance().getOrigin();
         // broadcast to peers
-        for (String p : peers) {
-            if (!p.equals(myOrigin)) {
-                EventEnvelope request = new EventEnvelope();
-                request.setTo(ServiceDiscovery.SERVICE_REGISTRY + "@" + p).setHeader(ORIGIN, myOrigin)
-                        .setHeader(TYPE, CHECKSUM).setHeader(CHECKSUM, checkSum);
-                po.send(request);
+        if (!peers.isEmpty()) {
+            List<String> myPeers = new ArrayList<>(peers);
+            myPeers.remove(origin);
+            if (myPeers.isEmpty()) {
+                log.info("No need to sync with peers because I am running alone");
+            } else {
+                log.info("Sending checksum {} to {}", checkSum, myPeers);
+                for (String p : myPeers) {
+                    EventEnvelope request = new EventEnvelope();
+                    request.setTo(ServiceDiscovery.SERVICE_REGISTRY + "@" + p).setHeader(ORIGIN, origin)
+                            .setHeader(TARGET, p).setHeader(TYPE, CHECKSUM).setHeader(CHECKSUM, checkSum);
+                    po.send(request);
+                }
             }
         }
     }