+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 1.12.62, 8/12/2020
+
+### Added
+
+New presence monitor's admin endpoint for the operator to force routing table synchronization ("/api/ping/now")
+
+### Removed
+
+N/A
+
+### Changed
+
+Improved routing table integrity check
+
+---
 ## Version 1.12.61, 8/8/2020
 
 ### Added